% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleg.R
\name{deleg}
\alias{deleg}
\title{delegate
This function delegate the execution of the passed command to iExec}
\usage{
deleg(cmd, encryption = T)
}
\arguments{
\item{cmd}{a command to execute}

\item{encryption}{a boolean indicating if encryption is required}
}
\value{
the function returns the id of the task created
}
\description{
delegate
This function delegate the execution of the passed command to iExec
}
