% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleg_caret.R
\name{deleg_caret}
\alias{deleg_caret}
\title{deleg_caret
This function delegate the execution of the passed command to iExec}
\usage{
deleg_caret(cmd, params = NULL, encryption = T, test = NULL)
}
\arguments{
\item{cmd}{a command to execute}

\item{params}{either a list of dataframe with the different values to have the function to execute}

\item{encryption}{a boolean indicating if encryption is required}

\item{test}{a dataset to test the created model against to evaluate it}
}
\value{
the function returns the ids of the deal and tasks created
}
\description{
deleg_caret
This function delegate the execution of the passed command to iExec
}
