% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResults.R
\name{getResults}
\alias{getResults}
\title{Title}
\usage{
getResults(taskId = NULL, jobId = NULL)
}
\arguments{
\item{taskId}{Id of the task to get the results of}

\item{jobId}{Id of the job to get the results of}
}
\description{
Title
}
